back-end/requirements.txt
click==8.0.3
dnspython==2.1.0
Flask==2.0.2
Flask-Cors==3.0.10
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
numpy==1.21.0
Pillow==9.0.0
pymongo==4.0.1
six==1.16.0
Werkzeug==2.0.2
